Add heartbeat thread per consumer
Rebalance of a group can cause a delay
in Consumer Heartbeat Thread and other
consumers may leave group if single
thread is used to iterate through
consumer states and send heartbeats
