+++ pom.xml
@@ -76,11 +76,6 @@
             <artifactId>kafka-json-serializer</artifactId>
             <version>${confluent.version}</version>
         </dependency>
-        <dependency>
-            <groupId>com.google.guava</groupId>
-            <artifactId>guava</artifactId>
-            <version>${guava.version}</version>
-        </dependency>
 
         <dependency>
             <groupId>io.confluent</groupId>
